added Rotation Vector Sensor and disabled
the gyroscope only sensor in favor of the
more meaningful rotation vector sensor.

Sensor changes the color to yellow when orientation is > 45 degrees
around Z-axis.

Sensor changes the color to blue when orientation is less than -45
degrees around Z-Axis.

Sensor changes the color to white when orientation is between 45
and 30 on the negative or positive scale around Z-axis.

Sensor does not change any color below absolute 30 around Z-axis
and does not log any events.